Defer resource downloading until needed
Improves startup time by not downloading resources during init, as the init function initializes _all_ optimizers, regardless if it's being used or not.

A `download` function has been added to classes that need this functionality for convenience.

Following the behavior from other classes with the same functionality, resources are only downloaded when an exception has been thrown, or the download function is called explicitly.